voice_control/requirements.txt
# Speech recognition
SpeechRecognition==3.10.0
pyaudio==0.2.13

# LangChain and LLM
langchain==0.1.4
langchain-google-genai==0.0.7
google-generativeai==0.3.1
pydantic==2.5.2

# MQTT Client
paho-mqtt==2.2.1

# Utilities
python-dotenv==1.0.0 